requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
jinja2==3.1.2
python-multipart==0.0.6
pyserial==3.5
requests==2.31.0
python-dotenv==1.0.0
